bugfix: save stars in XML, FilterView

+++ src/main/java/at/jku/se/diary/model/StructInformation.java
@@ -38,7 +38,7 @@ public class StructInformation {
         return stars;
     }
 
-    public void setStars(int stars) {
+    public void setStars(double stars) {
         this.stars = stars;
     }
